requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.4
Django==4.2.6
pycparser==2.21
PyMySQL==1.1.0
sqlparse==0.4.4
tzdata==2023.3
django-rest-framework